+++ src/integrated/soil_energy_hydrology_biogeochemistry.jl
@@ -57,7 +57,8 @@ end
     LandSoilBiogeochemistry{FT}(
         forcing,
         toml_dict::CP.ParamDict,
-        domain::Union{ClimaLand.Domains.Column, ClimaLand.Domains.SphericalShell};
+        domain::Union{ClimaLand.Domains.Column, ClimaLand.Domains.SphericalShell},
+        Δt;
         soil = Soil.EnergyHydrology{FT}(
             domain,
             forcing,
@@ -70,6 +71,7 @@ end
                 forcing.atmos,
             ),
             toml_dict,
+            Δt,
         ),
     ) where {FT}
 
@@ -79,6 +81,7 @@ or passed in via the `toml_dict`. The boundary conditions of all models
 correspond to `forcing` with the atmosphere, as specified by `forcing`, a NamedTuple
 of the form `(;atmos, radiation)`, with `atmos` an `AbstractAtmosphericDriver` and `radiation`
 an `AbstractRadiativeDriver`. The domain must be a ClimaLand domain with a vertical extent.
+`Δt` is the model timestep in seconds, used by the SoilCO2Model O2_f tendency limiter.
 """
 function LandSoilBiogeochemistry{FT}(
     forcing,
@@ -87,7 +90,8 @@ function LandSoilBiogeochemistry{FT}(
         ClimaLand.Domains.Column,
         ClimaLand.Domains.SphericalShell,
         ClimaLand.Domains.HybridBox,
-    };
+    },
+    Δt;
     soil = Soil.EnergyHydrology{FT}(domain, forcing, toml_dict;),
     soilco2 = Soil.Biogeochemistry.SoilCO2Model{FT}(
         domain,
@@ -96,6 +100,7 @@ function LandSoilBiogeochemistry{FT}(
             forcing.atmos,
         ),
         toml_dict,
+        Δt,
     ),
 ) where {FT}
     return LandSoilBiogeochemistry{FT}(soil, soilco2)
@@ -154,6 +159,15 @@ model from the prognostic liquid and ice fractions.
 """
 function soil_moisture(driver::PrognosticMet, p, Y, t, z)
     return p.soil.θ_l
+end
+
+"""
+    soil_ice(driver::PrognosticMet, p, Y, t, z)
+
+Returns the prognostic ice content from coupled soil model.
+"""
+function soil_ice(driver::PrognosticMet, p, Y, t, z)
+    return Y.soil.θ_i
 end